+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Enhancements
 
 - Add `/trace` endpoint [402](https://github.com/bugsnag/maze-runner/pull/402)
+- Add steps to set sampling probably response header [419](https://github.com/bugsnag/maze-runner/pull/419)
 
 
 ##############################